Refactor UI and update data models
Refactored `App.xaml` and `MainWindow.xaml` to improve UI structure and styling. Updated `MainWindow.xaml.cs` to remove unused methods and set `DataContext` to `ProductViewModel`. Changed data types in `BOM_Material.cs`, `Material.cs`, and `Product.cs` from `float` to `double` for better precision. Updated `ProductManagementApp.csproj` to manage dependencies and project structure. Added `BaseViewModel.cs` and `ProductViewModel.cs` to implement MVVM pattern.

+++ ProductManagementApp/Models/BOM_Material.cs
@@ -4,7 +4,7 @@
     {
         public int Material_Number { get; set; }
         public int BOM_Id { get; set; }
-        public float Qty { get; set; }
+        public double Qty { get; set; }             // Schimbat din float în double
         public string Unit_Measure_Code { get; set; }
     }
 }
