+++ ProductManagementApp/Models/Material.cs
@@ -4,8 +4,8 @@
     {
         public int Material_Number { get; set; }
         public string Material_Description { get; set; }
-        public float Weight { get; set; }
-        public float Width { get; set; }
-        public float Height { get; set; }
+        public double Weight { get; set; }  // Schimbat din float în double
+        public double Width { get; set; }   // Schimbat din float în double
+        public double Height { get; set; }  // Schimbat din float în double
     }
 }
