+++ ProductManagementApp/Models/Product.cs
@@ -5,9 +5,9 @@
         public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public float Estimated_Height { get; set; }
-        public float Estimated_Width { get; set; }
-        public float Estimated_Weight { get; set; }
+        public double Estimated_Height { get; set; } // Schimbat din float în double
+        public double Estimated_Width { get; set; }  // Schimbat din float în double
+        public double Estimated_Weight { get; set; } // Schimbat din float în double
         public int BOM_Id { get; set; }
     }
 }
